Realizado mudanças no ItemInventario pois o mesmo estava com erros. Adiiocnado os metodos de adicionar e remover quantidade do item, respeitando o mesmo para não aparecer numeros negativos

+++ src/main/java/com/senac/ControlaStock/repository/ItemInventarioRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface
-ItemInventarioRepository extends JpaRepository<ItemInventario, Integer> {
+ItemInventarioRepository extends JpaRepository<ItemInventario, Long> {
 
     Optional<ItemInventario> findByNome(String nome);
 }
